+++ app/models/csv_importer.rb
@@ -1,0 +1,17 @@
+require "csv"
+
+class CsvImporter
+  attr_reader :filepath, :formatter
+
+  def initialize(filepath, formatter)
+    @filepath = filepath
+    @formatter = formatter
+  end
+
+  def import
+    CSV.foreach(filepath, headers: true) do |row|
+      formatted_data = formatter.build_csv(row)
+      Product.create(formatted_data)
+    end
+  end
+end
